Fix JS unit tests
